+++ framework/src/modules/interoperability/base_cc_commands/channel_terminated.ts
@@ -12,10 +12,10 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-import { BaseInteroperabilityCCCommand } from '../../base_interoperability_cc_commands';
-import { channelTerminatedCCMParamsSchema } from '../../schemas';
-import { CROSS_CHAIN_COMMAND_NAME_CHANNEL_TERMINATED } from '../../constants';
-import { CCCommandExecuteContext } from '../../types';
+import { BaseInteroperabilityCCCommand } from '../base_interoperability_cc_commands';
+import { channelTerminatedCCMParamsSchema } from '../schemas';
+import { CROSS_CHAIN_COMMAND_NAME_CHANNEL_TERMINATED } from '../constants';
+import { CCCommandExecuteContext } from '../types';
 
 // LIP-0049 https://github.com/LiskHQ/lips/blob/main/proposals/lip-0049.md#channel-terminated-message-1
 export abstract class BaseCCChannelTerminatedCommand extends BaseInteroperabilityCCCommand {
